requirements.txt
Django==1.3.1
MySQL-python==1.2.3
PIL==1.1.7
South==0.7.6
pygeoip==0.2.4
six==1.2.0
wsgiref==0.1.2
django-livesettings
django-keyedcache
